Move util into src
