+++ src/Button/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import withStyles from '../../util/withStyles';
+import withStyles from '../util/withStyles';
 import styles from './index.scss';
 
 const Button = ({ onClick, label, className, disabled = false, children }) => {
